+++ modules/servers/apnscp/apnscp.php
@@ -8,13 +8,13 @@
  */
 
 use WHMCS\Database\Capsule as DB;
+use WHMCS\Service\Status;
 
-if (! defined("WHMCS"))
-{
+if (! defined("WHMCS")) {
     die("This file cannot be accessed directly");
 }
 
-require_once('lib/Connector.php');
+require_once('lib/ApisConnector.php');
 require_once('lib/Helper.php');
 
 /**
@@ -29,15 +29,21 @@ require_once('lib/Helper.php');
  */
 function apnscp_MetaData()
 {
-    return array(
+    return [
         'DisplayName'              => 'apnscp',
         'APIVersion'               => '1.0', // Use API Version 1.1
         'RequiresServer'           => true, // Set true if module requires a server to work
         'DefaultNonSSLPort'        => '2082', // Default Non-SSL Connection Port
         'DefaultSSLPort'           => '2083', // Default SSL Connection Port
-        'ServiceSingleSignOnLabel' => 'Login to Panel',
-        'AdminSingleSignOnLabel'   => 'Login to Panel as Admin',
-    );
+        'ServiceSingleSignOnLabel' => 'Login to ApisCP',
+        //        'AdminSingleSignOnLabel'                  => 'Login to ApisCP as Admin',
+        //        // The display name of the unique identifier to be displayed on the table output
+        //        'ListAccountsUniqueIdentifierDisplayName' => 'Domain',
+        //        // The field in the return that matches the unique identifier
+        //        'ListAccountsUniqueIdentifierField'       => 'domain',
+        //        // The config option indexed field from the _ConfigOptions function that identifies the product on the remote system
+        //        'ListAccountsProductField'                => 'configoption1',
+    ];
 }
 
 /**
@@ -65,10 +71,8 @@ function apnscp_MetaData()
  */
 function apnscp_ConfigOptions()
 {
-    $plans = apnscp_getPlans();
-
     return [
-        'apnscp Plan' => [
+        'ApisCP Plan' => [
             'Type'        => 'dropdown',
             'Default'     => 'basic',
             'Description' => 'Choose a plan (auto populated)<br>Format: &lt;server&gt; - &lt;plan&gt;',
@@ -90,15 +94,15 @@ function apnscp_ConfigOptions()
  *
  * @param array $params common module parameters
  *
+ * @return string "success" or an error message
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return string "success" or an error message
  */
 function apnscp_CreateAccount(array $params)
 {
     // Setup Server Params
     $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-    $apnscp_apikey      = $params['serverpassword'];
+    $apnscp_apikey = $params['serverpassword'];
 
     $domain = strtolower($params['domain']);
 
@@ -106,24 +110,19 @@ function apnscp_CreateAccount(array $params)
 
     $cliCommand = Helper::generateCommand($opts, 'AddDomain');
 
-    logModuleCall('apnscp', __FUNCTION__, ['CommandString' => $cliCommand], '', '');
-
-    try
-    {
-        $adminId = \session_id();
-        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
-
+    try {
+        $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
         $client->admin_add_site($domain, $params['username'], $opts);
-    }
-    catch (Exception $e)
-    {
+        Helper::apnscpValidateCustomFields($params['pid']);
+
+        logModuleCall('apnscp', 'Create', ['Request' => str_ireplace('><', ">\n<", $client->__getLastRequest()), 'CommandString' => $cliCommand], str_ireplace('><', ">\n<", $client->__getLastResponse()));
+    } catch (Exception $e) {
         // Record the error in WHMCS's module log.
         logModuleCall(
             'apnscp',
-            __FUNCTION__,
-            ['params' => $params, 'options' => $opts, 'CommandString' => $cliCommand],
-            $e->getMessage(),
-            $e->getTraceAsString()
+            'Create',
+            str_ireplace('><', ">\n<", $client->__getLastRequest()),
+            $e->getMessage() . "\n\n" . $e->getTraceAsString() . "\n\n" . str_ireplace('><', ">\n<", $client->__getLastResponse())
         );
 
         return $e->getMessage();
@@ -141,34 +140,31 @@ function apnscp_CreateAccount(array $params)
  *
  * @param array $params common module parameters
  *
+ * @return string "success" or an error message
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return string "success" or an error message
  */
 function apnscp_SuspendAccount(array $params)
 {
     $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-    $apnscp_apikey      = $params['serverpassword'];
-    $site_domain        = strtolower($params['domain']);
+    $apnscp_apikey = $params['serverpassword'];
+    $site_domain = strtolower($params['domain']);
 
     $opts['reason'] = $params['suspendreason'];
 
-    try
-    {
-        $adminId = \session_id();
-        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+    try {
+        $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
 
         $client->admin_deactivate_site($site_domain, $opts);
-    }
-    catch (Exception $e)
-    {
+
+        logModuleCall('apnscp', 'Suspend', ['Request' => str_ireplace('><', ">\n<", $client->__getLastRequest())], str_ireplace('><', ">\n<", $client->__getLastResponse()));
+    } catch (Exception $e) {
         // Record the error in WHMCS's module log.
         logModuleCall(
             'apnscp',
-            __FUNCTION__,
-            $params,
-            $e->getMessage(),
-            $e->getTraceAsString()
+            'Suspend',
+            str_ireplace('><', ">\n<", $client->__getLastRequest()),
+            $e->getMessage() . "\n\n" . $e->getTraceAsString() . "\n\n" . str_ireplace('><', ">\n<", $client->__getLastResponse())
         );
 
         return $e->getMessage();
@@ -186,32 +182,29 @@ function apnscp_SuspendAccount(array $params)
  *
  * @param array $params common module parameters
  *
+ * @return string "success" or an error message
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return string "success" or an error message
  */
 function apnscp_UnsuspendAccount(array $params)
 {
     $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-    $apnscp_apikey      = $params['serverpassword'];
-    $site_domain        = strtolower($params['domain']);
+    $apnscp_apikey = $params['serverpassword'];
+    $site_domain = strtolower($params['domain']);
 
-    try
-    {
-        $adminId = \session_id();
-        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+    try {
+        $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
 
         $client->admin_activate_site($site_domain);
-    }
-    catch (Exception $e)
-    {
+
+        logModuleCall('apnscp', 'Unsuspend', ['Request' => str_ireplace('><', ">\n<", $client->__getLastRequest())], str_ireplace('><', ">\n<", $client->__getLastResponse()));
+    } catch (Exception $e) {
         // Record the error in WHMCS's module log.
         logModuleCall(
             'apnscp',
-            __FUNCTION__,
-            $params,
-            $e->getMessage(),
-            $e->getTraceAsString()
+            'Unsuspend',
+            str_ireplace('><', ">\n<", $client->__getLastRequest()),
+            $e->getMessage() . "\n\n" . $e->getTraceAsString() . "\n\n" . str_ireplace('><', ">\n<", $client->__getLastResponse())
         );
 
         return $e->getMessage();
@@ -228,48 +221,62 @@ function apnscp_UnsuspendAccount(array $params)
  *
  * @param array $params common module parameters
  *
+ * @return string "success" or an error message
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return string "success" or an error message
  */
 function apnscp_TerminateAccount(array $params)
 {
+    // Logging
+    $module = 'apnscp';
+    $method = 'terminate';
+    // end
+
     $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-    $apnscp_apikey      = $params['serverpassword'];
-    $site_domain        = strtolower($params['domain']);
+    $apnscp_apikey = $params['serverpassword'];
+    $site_domain = strtolower($params['domain']);
 
-    try
-    {
-        $adminId = \session_id();
-        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+    try {
+        $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
 
-        /*
-         * To disable cancellation hold, uncomment these lines
-         */
-        $opts['force'] = 'true';
+        $opts['force'] = true;
         $client->admin_delete_site($site_domain, $opts);
-############################################################################################
-        /*
-         * To enable cancellation hold, uncomment these lines
-         */
+
 //        $opts['reason'] = 'Customer Requested Cancellation';
 //        $client->admin_deactivate_site($site_domain, $opts);
-    }
-    catch (Exception $e)
-    {
-        // Record the error in WHMCS's module log.
+
         logModuleCall(
-            'apnscp',
-            __FUNCTION__,
-            $params,
-            $e->getMessage(),
-            $e->getTraceAsString()
+            $module,
+            $method,
+            "Request: " . str_ireplace('><', ">\n<", $client->__getLastRequest()) . "\n\nHeaders:" . $client->__getLastRequestHeaders(),
+            "Response: " . $client->__getLastResponse() . "\n\n" .
+            "Headers: " . $client->__getLastResponseHeaders()
         );
+
+        return 'success';
+    } catch (Exception $e) {
+        logModuleCall(
+            $module,
+            $method,
+            "Request: " . str_ireplace('><', ">\n<", $client->__getLastRequest()) . "\n\nHeaders:" . $client->__getLastRequestHeaders(),
+            "Exception: " . $e->getMessage() . "\n\n" .
+            $e->getLine() . "\n\n" .
+            $e->getTraceAsString() . "\n\n" .
+            "Response: " . $client->__getLastResponse() . "\n\n" .
+            "Headers: " . $client->__getLastResponseHeaders()
+        );
+
+        if (empty($e->getMessage()) || ($e->getMessage() === 'Error Fetching http headers')) {
+            /**
+             * This is almost certainly a timeout error where the panel received and processed the termination command but didn't reply in time and SOAP threw an error
+             * If I'm wrong, I'll re-think this approach :)
+             */
+
+            return 'success';
+        }
 
         return $e->getMessage();
     }
-
-    return 'success';
 }
 
 /**
@@ -284,34 +291,31 @@ function apnscp_TerminateAccount(array $params)
  *
  * @param array $params common module parameters
  *
+ * @return string "success" or an error message
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return string "success" or an error message
  */
 function apnscp_ChangePassword(array $params)
 {
     $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-    $apnscp_apikey      = $params['serverpassword'];
-    $site_domain        = strtolower($params['domain']);
-    $site_admin         = $params['username'];
-    $site_password      = $params['password'];
+    $apnscp_apikey = $params['serverpassword'];
+    $site_domain = strtolower($params['domain']);
+    $site_admin = $params['username'];
+    $site_password = $params['password'];
 
-    try
-    {
-        $adminId = \session_id();
-        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+    try {
+        $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
 
         $client->auth_change_password($site_password, $site_admin, $site_domain);
-    }
-    catch (Exception $e)
-    {
+
+        logModuleCall('apnscp', 'ChangePassword', ['Request' => str_ireplace('><', ">\n<", $client->__getLastRequest())], str_ireplace('><', ">\n<", $client->__getLastResponse()));
+    } catch (Exception $e) {
         // Record the error in WHMCS's module log.
         logModuleCall(
             'apnscp',
-            __FUNCTION__,
-            $params,
-            $e->getMessage(),
-            $e->getTraceAsString()
+            'ChangePassword',
+            str_ireplace('><', ">\n<", $client->__getLastRequest()),
+            $e->getMessage() . "\n\n" . $e->getTraceAsString() . "\n\n" . $client->__getLastResponseHeaders() . "\n\n" . str_ireplace('><', ">\n<", $client->__getLastResponse())
         );
 
         return $e->getMessage();
@@ -332,41 +336,36 @@ function apnscp_ChangePassword(array $params)
  *
  * @param array $params common module parameters
  *
+ * @return string "success" or an error message
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return string "success" or an error message
  */
 function apnscp_ChangePackage(array $params)
 {
     // Setup Server Params
     $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-    $apnscp_apikey      = $params['serverpassword'];
+    $apnscp_apikey = $params['serverpassword'];
 
     $domain = strtolower($params['domain']);
 
     $opts['siteinfo.plan'] = $params['configoption1'];
-    $extra['reset'] = 'true';
+    $extra['reset'] = true;
 
     $cliCommand = Helper::generateCommand($opts, 'EditDomain');
 
-    logModuleCall('apnscp', __FUNCTION__, ['CommandString' => $cliCommand], '', '');
-
-    try
-    {
-        $adminId = \session_id();
-        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+    try {
+        $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
 
         $client->admin_edit_site($domain, $opts, $extra);
-    }
-    catch (Exception $e)
-    {
+
+        logModuleCall('apnscp', 'ChangePackage', ['Request' => str_ireplace('><', ">\n<", $client->__getLastRequest()), 'CommandString' => $cliCommand], str_ireplace('><', ">\n<", $client->__getLastResponse()));
+    } catch (Exception $e) {
         // Record the error in WHMCS's module log.
         logModuleCall(
             'apnscp',
-            __FUNCTION__,
-            ['params' => $params, 'options' => $opts],
-            $e->getMessage(),
-            $e->getTraceAsString()
+            'ChangePackage',
+            str_ireplace('><', ">\n<", $client->__getLastRequest()),
+            $e->getMessage() . "\n\n" . $e->getTraceAsString() . "\n\n" . str_ireplace('><', ">\n<", $client->__getLastResponse())
         );
 
         return $e->getMessage();
@@ -384,19 +383,19 @@ function apnscp_ChangePackage(array $params)
  *
  * @param array $params common module parameters
  *
+ * @return array
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return array
  */
 function apnscp_ServiceSingleSignOn(array $params)
 {
     $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-    $apnscp_apikey      = $params['serverpassword'];
-    $site_domain        = strtolower($params['domain']);
-    $site_admin         = $params['username'];
-    $app                = App::get_req_var('app');
-    $extra              = [];
-    $allowed_apps       = [
+    $apnscp_apikey = $params['serverpassword'];
+    $site_domain = strtolower($params['domain']);
+    $site_admin = $params['username'];
+    $app = App::get_req_var('app');
+    $extra = [];
+    $allowed_apps = [
         'usermanage',
         'mailboxroutes',
         'vacation',
@@ -412,25 +411,21 @@ function apnscp_ServiceSingleSignOn(array $params)
         'whitelist',
     ];
 
-    try
-    {
-        $adminId = \session_id();
-        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+    try {
+        $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
 
         $session_id = $client->admin_hijack($site_domain, $site_admin, 'UI');
 
-        if (! isset($app) OR ! in_array($app, $allowed_apps))
-        {
+        if (! isset($app) or ! in_array($app, $allowed_apps)) {
             $app = 'dashboard';
         }
 
-        if ($app === 'subdomains')
-        {
+        if ($app === 'subdomains') {
             $extra['mode'] = 'add';
         }
 
         $extra['esprit_id'] = $session_id;
-        $query              = http_build_query($extra);
+        $query = http_build_query($extra);
 
         $url = "${apnscp_apiendpoint}/apps/${app}?${query}";
 
@@ -438,22 +433,19 @@ function apnscp_ServiceSingleSignOn(array $params)
             'success'    => true,
             'redirectTo' => $url,
         ];
-    }
-    catch (Exception $e)
-    {
+    } catch (Exception $e) {
         // Record the error in WHMCS's module log.
         logModuleCall(
             'apnscp',
-            __FUNCTION__,
-            $params,
-            $e->getMessage(),
-            $e->getTraceAsString()
+            'SSO',
+            str_ireplace('><', ">\n<", $client->__getLastRequest()),
+            $e->getMessage() . "\n\n" . $e->getTraceAsString() . "\n\n" . str_ireplace('><', ">\n<", $client->__getLastResponse())
         );
 
-        return array(
+        return [
             'success'  => false,
             'errorMsg' => $e->getMessage(),
-        );
+        ];
     }
 }
 
@@ -467,9 +459,9 @@ function apnscp_ServiceSingleSignOn(array $params)
  *
  * @param array $params common module parameters
  *
+ * @return array
  * @see https://developers.whmcs.com/provisioning-modules/module-parameters/
  *
- * @return array
  */
 function apnscp_ClientArea(array $params)
 {
@@ -482,72 +474,77 @@ function apnscp_ClientArea(array $params)
 function apnscp_getPlans()
 {
     $servers = DB::table('tblservers')->where('type', 'apnscp')->get();
-    $adminId = \session_id();
 
-    try
-    {
-        foreach ($servers as $server)
-        {
+    try {
+        foreach ($servers as $server) {
             $apnscp_apiendpoint = ($server->secure === 'on' ? 'https' : 'http') . '://' . $server->hostname . ':' . ($server->secure === 'on' ? '2083' : '2082');
-            $apnscp_apikey      = decrypt($server->password);
+            $apnscp_apikey = decrypt($server->password);
 
-            $client = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+            $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
 
             $plans = $client->admin_list_plans();
 
-            foreach ($plans as $plan)
-            {
-                $return[ $plan ] = $server->name . ' - ' . $plan;
+            foreach ($plans as $plan) {
+                $return[$plan] = $server->name . ' - ' . $plan;
             }
         }
 
         return $return;
-    }
-    catch (Exception $e)
-    {
+    } catch (Exception $e) {
         // No easy way to return an error so we'll default to the basic plan only
         return ['basic' => 'basic (api call failed)'];
     }
 }
 
+function apnscp_UsageUpdate($params)
+{
+    $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
+    $apnscp_apikey = $params['serverpassword'];
+    $serverid = $params['serverid'];
+
+    $client = ApisConnector::create_client($apnscp_apikey, $apnscp_apiendpoint);
+    $siteInfoArr = $client->admin_collect(['siteinfo.domain']);
+    $storageArr = $client->admin_get_usage('storage');
+    $bandwidthArr = $client->admin_get_usage('bandwidth');
+
+    $products = DB::table('tblproducts')->where('type', 'hostingaccount')->where('servertype', 'apnscp')->get();
+    foreach ($products as $product) {
+        // This is less than efficient, but it works and is only a few extra queries!
+        Helper::apnscpValidateCustomFields($product->id);
+    }
+
+    foreach ($siteInfoArr as $site => $values) {
+        $service = DB::table('tblhosting')
+            ->where('server', $serverid)
+            ->where('domain', $values['domain'])
+            ->first();
+
+        if (! empty($service)) {
+            DB::table('tblhosting')
+                ->where('id', $service->id)
+                ->update([
+                    'diskusage'  => $storageArr[$site]['qused'] / 1024,
+                    'disklimit'  => $storageArr[$site]['qhard'] / 1024,
+                    'bwusage'    => ($bandwidthArr[$site]['sum'] / 1024) / 1024,
+                    'bwlimit'    => ($bandwidthArr[$site]['threshold'] / 1024) / 1024,
+                    'lastupdate' => DB::raw('now()'),
+                ]);
+
+            $customFields = Helper::apnscpGetCustomFields($service->packageid);
+            Helper::apnscpAddCustomFieldValue($service->id, $customFields['SiteID']['id'], $site);
+        }
+    }
+}
+
 
 /**
- * This is a hot mess on servers with a large number of accounts.
- * Disabled Usage Stats, takes way too long to run.  Every single domain requires 3 API calls to Hijack, get BW and get Disk usage
- * Multiplied by hundreds of accounts per server and many servers, this takes too long.
- * Enable only if absolutely necessary.  Please feel free to update and submit a PR in the repo.  This definitely needs a rework and optimization!
+ * Log module call.
+ *
+ * @param string $module The name of the module
+ * @param string $action The name of the action being performed
+ * @param string|array $requestString The input parameters for the API call
+ * @param string|array $responseData The response data from the API call
+ * @param string|array $processedData The resulting data after any post processing (eg. json decode, xml decode, etc...)
+ * @param array $replaceVars An array of strings for replacement
  */
-
-//function apnscp_UsageUpdate($params)
-//{
-//    $apnscp_apiendpoint = $params['serverhttpprefix'] . '://' . $params['serverhostname'] . ':' . $params['serverport'];
-//    $apnscp_apikey      = $params['serverpassword'];
-//    $serverid           = $params['serverid'];
-//    $stats              = [];
-//
-//    # Run connection to retrieve usage for all domains/accounts on $serverid
-//    $adminId     = \session_id();
-//    $adminClient = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
-//    $domains     = $adminClient->admin_get_domains();
-//    foreach ($domains as $domain)
-//    {
-//        $session_id = $adminClient->admin_hijack($domain);
-//        $userClient = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $session_id);
-//        $bw         = $userClient->site_get_bandwidth_usage();
-//        $quota      = $userClient->site_get_account_quota();
-//
-//        $stats[] = ['domain' => $domain, 'bw' => $bw, 'disk' => $quota];
-//    }
-//
-//    # Now loop through results and update DB
-//    foreach ($stats AS $values)
-//    {
-//        update_query('tblhosting', [
-//            'diskusage'  => $values['disk']['qused'] / 1024,
-//            'disklimit'  => $values['disk']['qhard'] / 1024,
-//            'bwusage'    => ($values['bw']['used'] / 1024) / 1024,
-//            'bwlimit'    => ($values['bw']['total'] / 1024) / 1024,
-//            'lastupdate' => 'now()',
-//        ], ['server' => $serverid, 'domain' => $values['domain']]);
-//    }
-//}
+//logModuleCall($module, $action, $requestString, $responseData, $processedData, $replaceVars);